[ElmSharp] Fix SVACE warnings
Layout.cs : the ClassName can be null. it should be checked before use.
Image.cs : the instance of "new MemoryStream()" should be disposed before end of function.
EvasImage.cs : the instance of "new MemoryStream()" should be disposed before end of function.

+++ src/ElmSharp/ElmSharp/Layout.cs
@@ -304,7 +304,7 @@ namespace ElmSharp
         {
             set
             {
-                if (value.IsDefault)
+                if (value.IsDefault && ClassName != null)
                 {
                     string part = ClassName.ToLower().Replace("elm_", "") + "/" + "bg";
                     EdjeObject.DeleteColorClass(part);
